Add files via upload
The connection between the server and the client is established, displaying an initial message for the client. The client has 5 attempts to guess the word. This is done by verifying the character entered by the client with all the characters of the chosen word, and if they match, the character '_' is replaced in the corresponding positions with the respective letter.